reacted to an number of issues reported by codebeat
Introduce source-HU to fine-Picking
https://github.com/metasfresh/metasfresh/issues/2298
