+++ de.metas.adempiere.adempiere/base/src/main/java/org/adempiere/ad/wrapper/POJOWrapper.java
@@ -566,20 +566,7 @@ public class POJOWrapper implements InvocationHandler, IInterfaceWrapper
 		}
 		else if (methodName.startsWith("set") && args.length == 1)
 		{
-			final String propertyNameLowerCase = methodName.substring(3);
-			final Class<?> paramType = method.getParameterTypes()[0];
-			final Object value = args[0];
-			if (isModelInterface(paramType))
-			{
-				setReferencedObject(propertyNameLowerCase, value);
-				// throw new AdempiereException("Object setter not supported: " + method);
-				// setValueFromPO(propertyName + "_ID", paramType, value);
-			}
-			else
-			{
-				setValue(propertyNameLowerCase, value);
-			}
-			return null;
+			return invokeSet(method, args, methodName);
 		}
 		else if (methodName.equals("get_TrxName") && (args == null || args.length == 0))
 		{
@@ -587,76 +574,11 @@ public class POJOWrapper implements InvocationHandler, IInterfaceWrapper
 		}
 		else if (methodName.startsWith("get") && (args == null || args.length == 0))
 		{
-			final String propertyNameLowerCase = methodName.substring(3);
-
-			if (propertyNameLowerCase.equals(idColumnName))
-			{
-				return getId();
-			}
-			if (isModelInterface(method.getReturnType()))
-			{
-				final Object referencedObject = getReferencedObject(propertyNameLowerCase, method);
-				if (referencedObject != null)
-				{
-					final String referencedObjectTrxName = getTrxName(referencedObject);
-					Check.assume(Objects.equals(this.getTrxName(), referencedObjectTrxName), "Invalid transaction"); // shall not happen, never ever
-					// POJOWrapper.setTrxName(referencedObject, POJOWrapper.getTrxName(this));
-				}
-				return referencedObject;
-			}
-
-			Object value = getValue(propertyNameLowerCase, method.getReturnType());
-			if (value != null)
-			{
-				return value;
-			}
-			//
-			if (method.getReturnType() == int.class)
-			{
-				if (propertyNameLowerCase.endsWith("_ID"))
-				{
-					value = DEFAULT_VALUE_ID;
-				}
-				else
-				{
-					value = DEFAULT_VALUE_int;
-				}
-			}
-			else if (method.getReturnType() == BigDecimal.class)
-			{
-				value = DEFAULT_VALUE_BigDecimal;
-			}
-			else if (PO.class.isAssignableFrom(method.getReturnType()))
-			{
-				throw new IllegalArgumentException("Method not supported - " + methodName);
-			}
-			return value;
+			return invokeGet(method, methodName);
 		}
 		else if (methodName.startsWith("is") && (args == null || args.length == 0))
 		{
-			final Map<String, Object> values = getInnerValues();
-			final String propertyNameLowerCase = methodName.substring(2);
-			final Object value;
-			if (values.containsKey(propertyNameLowerCase))
-			{
-				value = getValue(propertyNameLowerCase, method.getReturnType());
-			}
-			else if (values.containsKey("Is" + propertyNameLowerCase))
-			{
-				value = getValue("Is" + propertyNameLowerCase, method.getReturnType());
-			}
-			else
-			{
-				if (strictValues)
-				{
-					throw new IllegalStateException("No property " + propertyNameLowerCase + " was defined for bean " + this);
-				}
-
-				value = Boolean.FALSE;
-			}
-
-			// System.out.println("values="+values+", propertyName="+propertyName+" => value="+value);
-			return value == null ? false : (Boolean)value;
+			return invokeIs(method, methodName);
 		}
 		else if (methodName.equals("equals") && args.length == 1)
 		{
@@ -676,6 +598,99 @@ public class POJOWrapper implements InvocationHandler, IInterfaceWrapper
 		}
 	}
 
+	private Object invokeSet(final Method method, final Object[] args, final String methodName)
+	{
+		final String propertyNameLowerCase = methodName.substring(3);
+		final Class<?> paramType = method.getParameterTypes()[0];
+		final Object value = args[0];
+		if (isModelInterface(paramType))
+		{
+			setReferencedObject(propertyNameLowerCase, value);
+			// throw new AdempiereException("Object setter not supported: " + method);
+			// setValueFromPO(propertyName + "_ID", paramType, value);
+		}
+		else
+		{
+			setValue(propertyNameLowerCase, value);
+		}
+		return null;
+	}
+
+	private Object invokeGet(final Method method, final String methodName)
+	{
+		final String propertyNameLowerCase = methodName.substring(3);
+
+		if (propertyNameLowerCase.equals(idColumnName))
+		{
+			return getId();
+		}
+		if (isModelInterface(method.getReturnType()))
+		{
+			final Object referencedObject = getReferencedObject(propertyNameLowerCase, method);
+			if (referencedObject != null)
+			{
+				final String referencedObjectTrxName = getTrxName(referencedObject);
+				Check.assume(Objects.equals(this.getTrxName(), referencedObjectTrxName), "Invalid transaction"); // shall not happen, never ever
+				// POJOWrapper.setTrxName(referencedObject, POJOWrapper.getTrxName(this));
+			}
+			return referencedObject;
+		}
+
+		Object value = getValue(propertyNameLowerCase, method.getReturnType());
+		if (value != null)
+		{
+			return value;
+		}
+		//
+		if (method.getReturnType() == int.class)
+		{
+			if (propertyNameLowerCase.endsWith("_ID"))
+			{
+				value = DEFAULT_VALUE_ID;
+			}
+			else
+			{
+				value = DEFAULT_VALUE_int;
+			}
+		}
+		else if (method.getReturnType() == BigDecimal.class)
+		{
+			value = DEFAULT_VALUE_BigDecimal;
+		}
+		else if (PO.class.isAssignableFrom(method.getReturnType()))
+		{
+			throw new IllegalArgumentException("Method not supported - " + methodName);
+		}
+		return value;
+	}
+
+	private Object invokeIs(final Method method, final String methodName)
+	{
+		final Map<String, Object> values = getInnerValues();
+		final String propertyNameLowerCase = methodName.substring(2);
+		final Object value;
+		if (values.containsKey(propertyNameLowerCase))
+		{
+			value = getValue(propertyNameLowerCase, method.getReturnType());
+		}
+		else if (values.containsKey("Is" + propertyNameLowerCase))
+		{
+			value = getValue("Is" + propertyNameLowerCase, method.getReturnType());
+		}
+		else
+		{
+			if (strictValues)
+			{
+				throw new IllegalStateException("No property " + propertyNameLowerCase + " was defined for bean " + this);
+			}
+
+			value = Boolean.FALSE;
+		}
+
+		// System.out.println("values="+values+", propertyName="+propertyName+" => value="+value);
+		return value == null ? false : (Boolean)value;
+	}
+	
 	void setReferencedObject(final String propertyName, final Object value)
 	{
 		final String idPropertyName = propertyName + "_ID";
