+++ de.metas.handlingunits.base/src/main/java/de/metas/handlingunits/picking/SourceHUsRepository.java
@@ -1,6 +1,8 @@
 package de.metas.handlingunits.picking;
 
 import static org.adempiere.model.InterfaceWrapperHelper.load;
+import static org.adempiere.model.InterfaceWrapperHelper.newInstance;
+import static org.adempiere.model.InterfaceWrapperHelper.save;
 
 import java.util.Collection;
 import java.util.Comparator;
@@ -10,18 +12,22 @@ import java.util.List;
 import java.util.Set;
 import java.util.TreeSet;
 
+import org.adempiere.ad.dao.IQueryBL;
 import org.adempiere.util.Services;
+import org.slf4j.Logger;
 import org.springframework.stereotype.Service;
 
 import com.google.common.collect.ImmutableList;
 
 import de.metas.handlingunits.model.I_M_HU;
+import de.metas.handlingunits.model.I_M_Source_HU;
 import de.metas.handlingunits.picking.IHUPickingSlotBL.RetrieveActiveSourceHusQuery;
 import de.metas.handlingunits.trace.HUTraceEvent;
 import de.metas.handlingunits.trace.HUTraceRepository;
 import de.metas.handlingunits.trace.HUTraceSpecification;
 import de.metas.handlingunits.trace.HUTraceType;
 import de.metas.inoutcandidate.model.I_M_ShipmentSchedule;
+import de.metas.logging.LogManager;
 import lombok.NonNull;
 
 /*
@@ -54,6 +60,8 @@ import lombok.NonNull;
 @Service
 public class SourceHUsRepository
 {
+	private static final Logger logger = LogManager.getLogger(SourceHUsRepository.class);
+	
 	private final HUTraceRepository huTraceRepository;
 
 	private final PickingCandidateRepository pickingCandidateRepo;
@@ -144,4 +152,31 @@ public class SourceHUsRepository
 		}
 	}
 
+	public I_M_Source_HU addSourceHu(final int huId)
+	{
+		final I_M_Source_HU sourceHU = newInstance(I_M_Source_HU.class);
+		sourceHU.setM_HU_ID(huId);
+		save(sourceHU);
+
+		logger.info("Created one M_Source_HU record for M_HU_ID={}", huId);
+		return sourceHU;
+	}
+
+	/**
+	 * 
+	 * @param huId
+	 * @return {@code true} if anything was deleted.
+	 */
+	public boolean removeSourceHu(final int huId)
+	{
+		final IQueryBL queryBL = Services.get(IQueryBL.class);
+
+		final int deleteCount = queryBL.createQueryBuilder(I_M_Source_HU.class)
+				.addEqualsFilter(I_M_Source_HU.COLUMN_M_HU_ID, huId)
+				.create()
+				.delete();
+
+		return deleteCount > 0;
+	}
+	
 }
